requirements.txt
flask
pandas
scikit-learn
google-cloud-storage
supabase